pyproject.toml
[project]
name = "althea"
version = "0.1.0"
description = "GitHub-native ingress service for OpenClaw task dispatch"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.116.0",
  "uvicorn[standard]>=0.35.0",
  "httpx>=0.28.0",
  "pydantic>=2.11.0",
  "pydantic-settings>=2.10.0",
  "PyJWT>=2.10.0",
  "cryptography>=44.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.4.0",
  "pytest-asyncio>=1.1.0",
]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
